Moved alert tests to dir
