+++ tests/protocols/gmpv208/entities/alerts/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2018-2021 Greenbone Networks GmbH
+# Copyright (C) 2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -16,20 +16,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import RequiredArgument
-
-
-class GmpCloneAlertTestCase:
-    def test_clone(self):
-        self.gmp.clone_alert('a1')
-
-        self.connection.send.has_been_called_with(
-            '<create_alert>' '<copy>a1</copy>' '</create_alert>'
-        )
-
-    def test_missing_id(self):
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_alert('')
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.clone_alert(None)
+from .test_clone_alert import GmpCloneAlertTestMixin
+from .test_create_alert import GmpCreateAlertTestMixin
+from .test_delete_alert import GmpDeleteAlertTestMixin
+from .test_get_alert import GmpGetAlertTestMixin
+from .test_get_alerts import GmpGetAlertsTestMixin
+from .test_modify_alert import GmpModifyAlertTestMixin
+from .test_test_alert import GmpTestAlertTestMixin
+from .test_trigger_alert import GmpTriggerAlertTestMixin
